+++ scripts/item.js
@@ -17,14 +17,14 @@ export function createActorHeaderButton(config, buttons) {
         });
     }
 }
-async function actorConfig(actorDocument) {
-    if (!(actorDocument.type === 'character' || actorDocument.type === 'npc')) {
+async function actorConfig(actor) {
+    if (!(actor.type === 'character' || actor.type === 'npc')) {
         ui.notifications.info('This feature must be used on a character or npc!');
         return;
     }
     let selection = await chris.dialog('Apply all of Chris\'s automations to this actor?', [['Yes', true], ['No', false]]);
     if (!selection) return;
-    await game.modules.get('ddb-importer').api.chris.adjustActor(actorDocument);
+    await game.modules.get('ddb-importer').api.chris.adjustActor(actor);
     ui.notifications.info('Actor update complete!');
 }
 async function itemConfig(itemDocument) {
@@ -105,8 +105,9 @@ async function itemConfig(itemDocument) {
     let selection = await chris.dialog('Automation found, apply it? (' + foundCompendiumName + ')', options);
     if (!selection) return;
     ChatMessage.create({
-        speaker: {alias: name},
-        content: '<hr><b>' + compendiumItem.name + ':</b><br><hr>' + compendiumItem.system.description.value
+        'speaker': {alias: name},
+        'whisper': [game.user.id],
+        'content': '<hr><b>' + compendiumItem.name + ':</b><br><hr>' + compendiumItem.system.description.value
     });
     let originalItem = duplicate(itemDocument.toObject());
     originalItem.name = compendiumItem.name;